add base workerwrapper
